+++ Frends.AzureBlobStorage.ReadBlob/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.1.1] - 2024-01-31
+### Updated
+- Azure.Identity to version 1.10.4
+- Azure.Storage.Blobs to version 12.19.1
+
 ## [1.1.0] - 2022-12-16
 ### Added
 - OAuth2 as a new additional authentication method.